trying to fix networking for tms
