+++ Assets/Tilemaps/Scripts/Matrix.cs
@@ -23,7 +23,15 @@ namespace Tilemaps.Scripts
     {
         public static Matrix GetMatrix(MonoBehaviour behaviour)
         {
-            return behaviour.GetComponentInParent<Matrix>();
+            var matrix =  behaviour.GetComponentInParent<Matrix>();
+            
+            if (!matrix)
+            {    // TODO hacky way... won't work with multilevel and multi ships
+                var map = GameObject.FindGameObjectWithTag("Map");
+                matrix = map.GetComponentInChildren<Matrix>();
+            }
+
+            return matrix;
         }
 
         private MetaTileMap _metaTileMap;
