+++ Assets/Tilemaps/Scripts/Tiles/ConnectedTile.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using UnityEditor.Networking.PlayerConnection;
 using UnityEngine;
 using UnityEngine.Tilemaps;
