+++ Assets/Tilemaps/Scripts/Utils/RendererComparer.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿#if UNITY_EDITOR
+using System.Collections.Generic;
 using System.Reflection;
 using UnityEditorInternal;
 using UnityEngine;
@@ -32,3 +33,4 @@ namespace Tilemaps.Scripts.Utils
         }
     }
 }
+#endif
