+++ Assets/materials/Resources/OverlayCritHole.mat
@@ -23,8 +23,8 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     m_Floats:
-    - _Radius: 0.03
-    - _Shape: 2
+    - _Radius: 0.06
+    - _Shape: 1
     m_Colors:
     - _Center: {r: 0.5, g: 0.5, b: 0, a: 0}
     - _Color: {r: 1, g: 1, b: 1, a: 1}
