+++ Assets/scenes/DeathMatch.unity
@@ -75274,7 +75274,7 @@ Prefab:
     - target: {fileID: 114598896978409020, guid: bdc8c5eacef3f4b2a90445083973323d,
         type: 2}
       propertyPath: m_Enabled
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 114598896978409020, guid: bdc8c5eacef3f4b2a90445083973323d,
         type: 2}
