+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,18 +5,18 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
-    path: Assets/scenes/Lobby.unity
+  - enabled: 1
+    path: Assets/Scenes/Lobby.unity
     guid: ec6f1f009e7b4ebf902aba40b6f1eae6
   - enabled: 0
-    path: Assets/scenes/BoxStation.unity
+    path: Assets/Scenes/BoxStation.unity
     guid: 7af1f11ebc23aee4fb0e35bd3052d788
   - enabled: 0
-    path: Assets/scenes/TestMap.unity
+    path: Assets/Scenes/TestMap.unity
     guid: 6e3c9c4b7b5b248438dba6e22f0f82d4
-  - enabled: 0
-    path: Assets/scenes/DeathMatch.unity
-    guid: efae39b21fbe7453894657699ec02fa8
   - enabled: 1
-    path: Assets/scenes/Outpost.unity
+    path: Assets/Scenes/DeathMatch.unity
+    guid: efae39b21fbe7453894657699ec02fa8
+  - enabled: 0
+    path: Assets/Scenes/Outpost.unity
     guid: 718afb9f82f884bf286fe47c5d991f24
